add support of secondary index
With this patch both read and scan operations can utilize a secondary index.
Secondary index can be provided as a suffix to the table name, separated by a colon.
For other operations the suffix will be ignored.

In case of read operation without a specified index a GetItem request is used.
With this patch, if a secondary index is specified, a Query request will be used instead.
Using "empty" index name (just a colon at the end of table name) allows to force the Query request on the base table.

Scan operation directly supports secondary indexes.
However on DynamoDB's scan, startkey is "exclusive", while in general YCSB expects it to be included in the results.
In previous implementations the first key was read with GetItem request.
In this patch by default it is read following the same logic as above read operations.
Additionally, this patch adds an option to set dynamodb.inclusiveScan=false to disable this additional read.